# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "koora API",
        "contact": {
            "name": "Ahmed Magdy and abouzaid",
            "url": "https://www.google.com"
        },
        "version": "1.0"
    },
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/competitions": {
            "get": {
                "tags": [
                    "Competition"
                ],
                "description": "get all competitions",
                "operationId": "AllCompetitions",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Competition"
                ],
                "description": "Create new Competition",
                "operationId": "createCompetition",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/competition"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "SUCCESSFULLY CREATED"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/api/Seasons/{season}/stages/{stage}/groups/{group}/teams": {
            "post": {
                "tags": [
                    "groups"
                ],
                "description": "Add Teams to a group",
                "operationId": "addGroupTeams",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "description": "Season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "stage",
                        "in": "path",
                        "description": "stage ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "description": "group ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groupTeam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "SUCCESSFULLY CREATED"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/api/Players/All-Players": {
            "get": {
                "tags": [
                    "Player"
                ],
                "description": "get all Players",
                "operationId": "GET_ALL_Teams",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/api/Teams/All-Teams": {
            "get": {
                "tags": [
                    "Team"
                ],
                "description": "get all Teams",
                "operationId": "GET_ALL_Teams",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/Teams/{id}": {
            "get": {
                "tags": [
                    "Team"
                ],
                "description": "get Team with it's id",
                "operationId": "GET_Team",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/Teams/Create": {
            "post": {
                "tags": [
                    "Team"
                ],
                "description": "post Create Team form ",
                "operationId": "POST_Create_Team",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Team_creation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/Teams/Update/{id}": {
            "put": {
                "tags": [
                    "Team"
                ],
                "description": "post Create Team form ",
                "operationId": "POST_Create_Team",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "team id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Team_creation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/Teams/Delete/{id}": {
            "delete": {
                "tags": [
                    "Team"
                ],
                "summary": "Remove Team from database",
                "operationId": "deleteTeam",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "competition": {
            "properties": {
                "name": {
                    "description": "competition name",
                    "type": "string",
                    "example": "Egyptian league"
                },
                "type": {
                    "description": "competition type id",
                    "type": "integer",
                    "example": "1 for league & 2 for cup"
                },
                "country_id": {
                    "description": "competition country id",
                    "type": "integer",
                    "example": "(if the competition is local)"
                },
                "continent_id": {
                    "description": "competition continent id",
                    "type": "integer",
                    "example": "(if the competition is continental)"
                }
            }
        },
...
        "groupTeam": {
            "properties": {
                "team_id": {
                    "type": "string",
                    "example": 1
                }
            }
        },
...
        "Team_creation": {
            "properties": {
                "name": {
                    "type": "string",
                    "example": "banha"
                },
                "type_id": {
                    "type": "integer",
                    "example": "1"
                },
                "logo": {
                    "type": "string",
                    "example": "logo.png"
                },
                "stadium": {
                    "type": "string",
                    "example": "Banha stadium"
                },
                "country_id": {
                    "type": "integer",
                    "example": "2"
                }
            }
        },
...
    }
}